se ControlledHighTable
